Move get/setters from WMS to OGC to give access on WFS
