+++ service-capabilities/src/main/java/org/oskari/capabilities/ogc/LayerCapabilitiesWFS.java
@@ -1,0 +1,9 @@
+package org.oskari.capabilities.ogc;
+
+public class LayerCapabilitiesWFS extends LayerCapabilitiesOGC {
+
+    public LayerCapabilitiesWFS(String name, String title) {
+        super(name, title);
+    }
+
+}
